refactor: fix pr requested changes
